Instructions for using gist and hub
Instructions for using gist and hub with Cisco Github Enterprise. gist
works, I have it aliased as cist. Hub is not logging in properly to
Enterprise, I think it's a certificate issue with the way Go reads
certificates from the certificate store on OS X.

However, hub is working with regular Github.