Trying Perforce Merge
I am working on finding the right diff/merge tool on the mac. I tried
Beyond Compare, but it had some problems diffing a file where the only
change was uncommenting a bunch of lines. It did not handle the diff
properly, and it made it look like the lines had changed when the only
change was removing the comment character.

So, I installed meld (brew install homebrew/x11/meld) and perforce
merge. I edited my git config so that perforce merge is now the default.
We'll see how it goes.